+++ libs/qt/source/ftrack_qt/widgets/headers/accordion_header_widget.py
@@ -19,6 +19,8 @@ class AccordionHeaderWidget(QtWidgets.QFrame):
     clicked = QtCore.Signal(object)  # User header click
     arrow_clicked = QtCore.Signal(object)  # User header click
     checkbox_status_changed = QtCore.Signal(object)
+    show_options_overlay = QtCore.Signal(object)
+    hide_options_overlay = QtCore.Signal()
 
     @property
     def title(self):
@@ -43,6 +45,10 @@ class AccordionHeaderWidget(QtWidgets.QFrame):
     @property
     def collapsed(self):
         return self._collapsed
+
+    @property
+    def options_button(self):
+        return self._options_button
 
     def __init__(
         self,
@@ -138,6 +144,18 @@ class AccordionHeaderWidget(QtWidgets.QFrame):
     def post_build(self):
         self._checkbox.stateChanged.connect(self._on_checkbox_status_changed)
         self._arrow.clicked.connect(self._on_arrow_clicked)
+        self._options_button.show_overlay_signal.connect(
+            self.on_show_options_callback
+        )
+        self._options_button.hide_overlay_signal.connect(
+            self.on_hide_options_callback
+        )
+
+    def on_show_options_callback(self, widget):
+        self.show_options_overlay.emit(widget)
+
+    def on_hide_options_callback(self):
+        self.hide_options_overlay.emit()
 
     def add_option_widget(self, widget, section_name):
         self._options_button.add_widget(widget, section_name)
@@ -168,3 +186,8 @@ class AccordionHeaderWidget(QtWidgets.QFrame):
     def set_status(self, status, message):
         '''Set status message within header, to be implemented by child'''
         pass
+
+    def teardown(self):
+        '''Teardown the options button - properly cleanup the options overlay'''
+        self._options_button.teardown()
+        self._options_button.deleteLater()